requirements.txt
arrow==0.15.7
beautifulsoup4==4.9.0
dj-database-url==0.5.0
django==3.0.6
django-cors-headers==3.3.0
django-dotenv==1.4.2
django-redis==4.11.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
googlemaps==4.4.0
gunicorn==20.0.4
pandas
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pymongo==3.10.1
redis==3.5.2
requests==2.23.0
